Accepted as the correct answer

+++ C/Input Output/Problem N/solution.c
@@ -2,10 +2,10 @@
 
 int main(){ 
     char name1 [100], name2 [100];
-    float height1, height2; 
+    double height1, height2; 
     int age1, age2; 
 
-    scanf("%s %f %d %s %f %d", name1, &height1, &age1, name2, &height2, &age2);
+    scanf("%s %lf %d %s %lf %d", name1, &height1, &age1, name2, &height2, &age2);
     
     printf("Name 1: %s\n", name1);
     printf("Height 1: %.2f\n", height1);
